+++ src/plugins/index.ts
@@ -1,0 +1,1 @@
+export { default as breakpoints } from '@/plugins/Breakpoints/Breakpoints'