+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 - Update button styles - [#61](https://github.com/PrefectHQ/miter-design/pull/61)
+- Add placeholder prop to Input - [#70](https://github.com/PrefectHQ/miter-design/pull/70)
+- Add breakpoints plugin - [#68](https://github.com/PrefectHQ/miter-design/pull/68)
 - Update tab component to match new dashboard designs and remove icon but - [#63](https://github.com/PrefectHQ/miter-design/pull/63)
 - Add input component - [#40](https://github.com/PrefectHQ/miter-design/pull/40)
 - Enforce slot styling from card component - [#57](https://github.com/PrefectHQ/miter-design/pull/57)